requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pyjwt
bcrypt
pydantic
pytest
httpx